refactor: extract `getSelectedBlock` to clipboard
