+++ packages/editor/src/managers/clipboard/paste-manager.ts
@@ -3,8 +3,6 @@ import {
   deleteModelsByRange,
   getStartModelBySelection,
   handleBlockSplit,
-  SelectionInfo,
-  SelectionUtils,
 } from '@blocksuite/blocks';
 import { assertExists, matchFlavours } from '@blocksuite/global/utils';
 import { BaseBlockModel, Text } from '@blocksuite/store';
@@ -13,6 +11,7 @@ import type { DeltaOperation } from 'quill';
 import type { EditorContainer } from '../../components/index.js';
 import { MarkdownUtils } from './markdown-utils.js';
 import { CLIPBOARD_MIMETYPE } from './types.js';
+import { getSelectInfo } from './utils.js';
 
 export class PasteManager {
   private _editor: EditorContainer;
@@ -177,21 +176,17 @@ export class PasteManager {
   }
 
   // TODO Max 15 deeper
-  public insertBlocks(blocks: OpenBlockInfo[], selectInfo?: SelectionInfo) {
+  public insertBlocks(
+    blocks: OpenBlockInfo[],
+    selectionInfo = getSelectInfo(this._editor.page)
+  ) {
     if (blocks.length === 0) {
       return;
     }
 
-    const currentSelectionInfo =
-      selectInfo || SelectionUtils.getSelectInfo(this._editor.page);
-    if (
-      currentSelectionInfo.type === 'Range' ||
-      currentSelectionInfo.type === 'Caret'
-    ) {
+    if (selectionInfo.type === 'Range' || selectionInfo.type === 'Caret') {
       const lastBlock =
-        currentSelectionInfo.selectedBlocks[
-          currentSelectionInfo.selectedBlocks.length - 1
-        ];
+        selectionInfo.selectedBlocks[selectionInfo.selectedBlocks.length - 1];
       const selectedBlock = this._editor.page.getBlockById(lastBlock.id);
       let parent = selectedBlock;
       let index = 0;
@@ -275,13 +270,12 @@ export class PasteManager {
             this._addBlocks(blocks.slice(0, 1), parent, index, addBlockIds);
             this._editor.page.deleteBlock(lastBlockModel);
           } else {
-            if (currentSelectionInfo.type === 'Range') {
+            if (selectionInfo.type === 'Range') {
               const textLength = selectedBlock.text?.length as number;
               deleteModelsByRange(this._editor.page);
-              const startPos = currentSelectionInfo.selectedBlocks[0]
+              const startPos = selectionInfo.selectedBlocks[0]
                 .startPos as number;
-              const endPos = currentSelectionInfo.selectedBlocks[0]
-                .endPos as number;
+              const endPos = selectionInfo.selectedBlocks[0].endPos as number;
               if (startPos + endPos === textLength + 1) {
                 selectedBlock?.text?.insertList(insertTexts, startPos);
               } else {
@@ -332,11 +326,9 @@ export class PasteManager {
       } else {
         parent && this._addBlocks(blocks, parent, index, addBlockIds);
       }
-    } else if (currentSelectionInfo.type === 'Block') {
+    } else if (selectionInfo.type === 'Block') {
       const selectedBlock = this._editor.page.getBlockById(
-        currentSelectionInfo.selectedBlocks[
-          currentSelectionInfo.selectedBlocks.length - 1
-        ].id
+        selectionInfo.selectedBlocks[selectionInfo.selectedBlocks.length - 1].id
       );
 
       let parent = selectedBlock;
